+++ Assets/PlayMaker Vuforia/Changelog.md
@@ -1,11 +1,15 @@
 #PlayMaker Vuforia Change log
 
+###2.1.1
+**Release Date:** 17/06/2018  
+
+**new** Support for 7.2.20 Vuforia library, including legacy. It affects the action `VuforiaGetExtendedTracking`. Now we are supposed to use Vuforia's `Device Tracker`
+
+
 ###2.1.0
 **Release:** 18/07/2018  
 
-**Update:**
-
-- PlayMaker Utils Submodule
+**Update:** PlayMaker Utils Submodule
 
 
 ###2.0.0